requirements.txt
<<<<<<< HEAD
fastapi==0.135.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.48
psycopg2-binary
alembic==1.12.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.3.0
=======
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary
alembic==1.12.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
>>>>>>> cb8739c9cc3aa7a178da3c1967b6e0798d062a28
